harness: extract message result into its own mod
