+++ harness/src/lib.rs
@@ -451,6 +451,7 @@ pub mod file;
 #[cfg(any(feature = "fuzz", feature = "fuzz-fd"))]
 pub mod fuzz;
 pub mod instructions_sysvar;
+mod message_result;
 pub mod program;
 #[cfg(feature = "register-tracing")]
 pub mod register_tracing;
@@ -466,20 +467,18 @@ use mollusk_svm_result::Compare;
 use {
     crate::{
         account_store::AccountStore, callback::invoke_context::MolluskInvokeContextCallback,
-        epoch_stake::EpochStake, program::ProgramCache, sysvar::Sysvars,
+        epoch_stake::EpochStake, message_result::MessageResult, program::ProgramCache,
+        sysvar::Sysvars,
     },
     mollusk_svm_error::error::{MolluskError, MolluskPanic},
     mollusk_svm_result::{
-        types::{TransactionProgramResult, TransactionResult},
-        Check, CheckContext, Config, InstructionResult,
+        types::TransactionResult, Check, CheckContext, Config, InstructionResult,
     },
     solana_account::{Account, AccountSharedData, ReadableAccount},
     solana_compute_budget::compute_budget::ComputeBudget,
     solana_hash::Hash,
     solana_instruction::{AccountMeta, Instruction},
-    solana_instruction_error::InstructionError,
     solana_message::SanitizedMessage,
-    solana_program_error::ProgramError,
     solana_program_runtime::{
         invoke_context::{EnvironmentConfig, InvokeContext},
         loaded_programs::ProgramRuntimeEnvironments,
@@ -558,59 +557,6 @@ impl CheckContext for Mollusk {
     fn is_rent_exempt(&self, lamports: u64, space: usize, owner: Pubkey) -> bool {
         owner.eq(&Pubkey::default()) && lamports == 0
             || self.sysvars.rent.is_exempt(lamports, space)
-    }
-}
-
-struct MessageResult {
-    /// The number of compute units consumed by the transaction.
-    pub compute_units_consumed: u64,
-    /// The time taken to execute the transaction, in microseconds.
-    pub execution_time: u64,
-    /// The raw result of the transaction's execution.
-    pub raw_result: Result<(), TransactionError>,
-    /// The return data produced by the transaction, if any.
-    pub return_data: Vec<u8>,
-    /// Inner instructions (CPIs) invoked during the transaction execution.
-    ///
-    /// Each entry represents a cross-program invocation made by the program,
-    /// including the invoked instruction and the stack height at which it
-    /// was called.
-    #[cfg(feature = "inner-instructions")]
-    pub inner_instructions: Vec<Vec<InnerInstruction>>,
-    /// The compiled message used to execute the transaction.
-    ///
-    /// This can be used to map account indices in inner instructions back to
-    /// their corresponding pubkeys via `message.account_keys()`.
-    ///
-    /// This is `None` when the result is loaded from a fuzz fixture, since
-    /// fixtures don't contain the compiled message.
-    #[cfg(feature = "inner-instructions")]
-    pub message: Option<SanitizedMessage>,
-}
-
-impl MessageResult {
-    fn extract_ix_err(txn_err: TransactionError) -> InstructionError {
-        match txn_err {
-            TransactionError::InstructionError(_, ix_err) => ix_err,
-            _ => unreachable!(), // Mollusk only uses `InstructionError` variant.
-        }
-    }
-
-    fn extract_txn_program_result(
-        raw_result: &Result<(), TransactionError>,
-    ) -> TransactionProgramResult {
-        match raw_result {
-            Ok(()) => TransactionProgramResult::Success,
-            Err(TransactionError::InstructionError(idx, ix_err)) => {
-                let index = *idx as usize;
-                if let Ok(program_error) = ProgramError::try_from(ix_err.clone()) {
-                    TransactionProgramResult::Failure(index, program_error)
-                } else {
-                    TransactionProgramResult::UnknownError(index, ix_err.clone())
-                }
-            }
-            _ => unreachable!(), // Mollusk only uses `InstructionError` variant.
-        }
     }
 }
